feat: add helpers file in communication/ros2 tests directory
