+++ tests/communication/ros2/test_api.py
@@ -14,121 +14,26 @@
 
 import threading
 import time
-from typing import Generator, List, Tuple
 
 import pytest
-import rclpy
 from action_msgs.msg import GoalStatus
 from action_msgs.srv import CancelGoal
-from nav2_msgs.action import NavigateToPose
-from rclpy.action import ActionServer
-from rclpy.action.client import ClientGoalHandle
-from rclpy.executors import MultiThreadedExecutor, SingleThreadedExecutor
+from rclpy.executors import MultiThreadedExecutor
 from rclpy.node import Node
-from std_msgs.msg import String
-from std_srvs.srv import SetBool
 
 from rai.communication.ros2.api import ROS2ActionAPI, ROS2ServiceAPI, ROS2TopicAPI
 
+from .helpers import ActionServer_ as ActionServer
+from .helpers import (
+    MessagePublisher,
+    MessageReceiver,
+    ServiceServer,
+    ros_setup,
+    shutdown_executors_and_threads,
+    single_threaded_spinner,
+)
 
-class ServiceServer(Node):
-    def __init__(self, service_name: str):
-        super().__init__("test_service_server")
-        self.srv = self.create_service(SetBool, service_name, self.handle_test_service)
-
-    def handle_test_service(
-        self, request: SetBool.Request, response: SetBool.Response
-    ) -> SetBool.Response:
-        response.success = True
-        response.message = "Test service called"
-        return response
-
-
-class MessageReceiver(Node):
-    def __init__(self, topic: str):
-        super().__init__("test_message_receiver")
-        self.subscription = self.create_subscription(
-            String, topic, self.handle_test_message, 10
-        )
-        self.received_messages: List[String] = []
-
-    def handle_test_message(self, msg: String) -> None:
-        self.received_messages.append(msg)
-
-
-class ActionServer_(Node):
-    def __init__(self, action_name: str):
-        super().__init__("test_action_server")
-        self.action_server = ActionServer(
-            self,
-            action_type=NavigateToPose,
-            action_name=action_name,
-            execute_callback=self.handle_test_action,
-        )
-
-    def handle_test_action(
-        self, goal_handle: ClientGoalHandle
-    ) -> NavigateToPose.Result:
-
-        for i in range(1, 11):
-            feedback_msg = NavigateToPose.Feedback(distance_remaining=10.0 / i)
-            goal_handle.publish_feedback(feedback_msg)
-            time.sleep(0.01)
-
-        goal_handle.succeed()
-
-        result = NavigateToPose.Result()
-        result.error_code = NavigateToPose.Result.NONE
-        return result
-
-
-class MessagePublisher(Node):
-    def __init__(self, topic: str):
-        super().__init__("test_message_publisher")
-        self.publisher = self.create_publisher(String, topic, 10)
-        self.timer = self.create_timer(0.1, self.publish_message)
-
-    def publish_message(self) -> None:
-        msg = String()
-        msg.data = "Hello, ROS2!"
-        self.publisher.publish(msg)
-
-
-def single_threaded_spinner(
-    nodes: List[Node],
-) -> Tuple[List[SingleThreadedExecutor], List[threading.Thread]]:
-    executors: List[SingleThreadedExecutor] = []
-    executor_threads: List[threading.Thread] = []
-    for node in nodes:
-        executor = SingleThreadedExecutor()
-        executor.add_node(node)
-        executors.append(executor)
-    for executor in executors:
-        executor_thread = threading.Thread(target=executor.spin)
-        executor_thread.daemon = True
-        executor_thread.start()
-        executor_threads.append(executor_thread)
-    return executors, executor_threads
-
-
-def shutdown_executors_and_threads(
-    executors: List[SingleThreadedExecutor], threads: List[threading.Thread]
-) -> None:
-    # First shutdown executors
-    for executor in executors:
-        executor.shutdown()
-    # Small delay to allow executors to finish pending operations
-    time.sleep(0.5)
-    # Then join threads with a timeout
-    for thread in threads:
-        thread.join(timeout=2.0)
-
-
-@pytest.fixture(scope="function")
-def ros_setup() -> Generator[None, None, None]:
-    rclpy.init()
-    yield
-    rclpy.shutdown()
+ros_setup  # type: ignore prevents pytest from complaining about unused fixture
 
 
 def test_ros2_single_message_publish(
@@ -377,7 +282,7 @@ def test_ros2_service_single_call_wrong_service_name(
 def test_ros2_action_send_goal(ros_setup: None, request: pytest.FixtureRequest) -> None:
     action_name = f"{request.node.originalname}_action"  # type: ignore
     node_name = f"{request.node.originalname}_node"  # type: ignore
-    action_server = ActionServer_(action_name)
+    action_server = ActionServer(action_name)
     node = Node(node_name)
     executors, threads = single_threaded_spinner([action_server, node])
 
@@ -398,7 +303,7 @@ def test_ros2_action_send_goal_get_result(
 ) -> None:
     action_name = f"{request.node.originalname}_action"  # type: ignore
     node_name = f"{request.node.originalname}_node"  # type: ignore
-    action_server = ActionServer_(action_name)
+    action_server = ActionServer(action_name)
     node = Node(node_name)
     executors, threads = single_threaded_spinner([action_server, node])
 
@@ -429,7 +334,7 @@ def test_ros2_action_send_goal_wrong_action_type(
 ) -> None:
     action_name = f"{request.node.originalname}_action"  # type: ignore
     node_name = f"{request.node.originalname}_node"  # type: ignore
-    action_server = ActionServer_(action_name)
+    action_server = ActionServer(action_name)
     node = Node(node_name)
     executors, threads = single_threaded_spinner([action_server, node])
 
@@ -450,7 +355,7 @@ def test_ros2_action_send_goal_wrong_action_name(
 ) -> None:
     action_name = f"{request.node.originalname}_action"  # type: ignore
     node_name = f"{request.node.originalname}_node"  # type: ignore
-    action_server = ActionServer_(action_name)
+    action_server = ActionServer(action_name)
     node = Node(node_name)
     executors, threads = single_threaded_spinner([action_server, node])
 
@@ -470,7 +375,7 @@ def test_ros2_action_send_goal_get_feedback(
 ) -> None:
     action_name = f"{request.node.originalname}_action"  # type: ignore
     node_name = f"{request.node.originalname}_node"  # type: ignore
-    action_server = ActionServer_(action_name)
+    action_server = ActionServer(action_name)
     node = Node(node_name)
     executors, threads = single_threaded_spinner([action_server])
     executor = MultiThreadedExecutor()
@@ -498,7 +403,7 @@ def test_ros2_action_send_goal_terminate_goal(
 ) -> None:
     action_name = f"{request.node.originalname}_action"  # type: ignore
     node_name = f"{request.node.originalname}_node"  # type: ignore
-    action_server = ActionServer_(action_name)
+    action_server = ActionServer(action_name)
     node = Node(node_name)
     executors, threads = single_threaded_spinner([action_server])
     executor = MultiThreadedExecutor()
